Refactor views into nested views from `SafetyElementsView`
So then we can have smaller names that are just available for this sub component
