+++ Demo/Components/LinkButtonList/LinkButtonListDemoView.swift
@@ -72,4 +72,3 @@ private extension Array where Element == LinkButtonViewModel {
         ]
     }()
 }
-
