+++ Demo/Components/SafetyElementsView/SafetyElementsDemoView.swift
@@ -42,32 +42,34 @@ class SafetyElementsDemoView: UIView, Tweakable {
             title: "Service",
             icon: UIImage(named: "service")!,
             body: "Selger garanterer at bilen har blitt vedlikeholdt i tråd med fabrikkens anbefalinger for service-intervall og øvrig vedlikehold.",
-            externalLink: nil
+            bottomLink: nil
         ),
         SafetyElementViewModel(
             title: "Garanti",
             icon: UIImage(named: "warranty")!,
             body: "Denne bilen selges med garanti. Nybilgaranti fra 1. gang reg. (60 mnd / 100 000 km)",
-            externalLink: LinkButtonViewModel(
+            bottomLink: LinkButtonViewModel(
                 buttonIdentifier: nil,
                 buttonTitle: "Les mer om de forskjellige garantitypene på FINNs hjelpesenter.",
-                linkUrl: URL(string: "https://hjelpesenter.finn.no/hc/no/articles/115002091069-Nytting-info-om-garanti-til-deg-som-skal-handle-bil-på-FINN") ?? URL(string: "https://google.com")!
+                linkUrl: URL(string: "https://google.com")!,
+                isExternal: true
             )
         ),
         SafetyElementViewModel(
             title: "Bytterett",
             icon: UIImage(named: "warranty")!,
             body: "Denne bilen selges med bytterett fra forhandler. Det betyr at du som kjøper har mulighet til å bytte den mot en annen bil, i henhold til forhandlerens vilkår. Dette kan du gjøre dersom bilen du har kjøpt av en eller annen grunn ikke tilfredsstiller dine forventninger eller behov. Produktet tilbys for at du som kjøper skal føle deg helt trygg på at du ender opp med riktig bil for ditt behov.",
-            externalLink: nil
+            bottomLink: nil
         ),
         SafetyElementViewModel(
             title: "Medlem",
             icon: UIImage(named: "warranty")!,
             body: "Når du kjøper bil av en forhandler som er medlem av NBF, handler du av en forhandler som har forpliktet seg til å følge NBF etiske retningslinjer. Dette gir deg som kjøper en ekstra trygghet.",
-            externalLink: LinkButtonViewModel(
+            bottomLink: LinkButtonViewModel(
                 buttonIdentifier: nil,
                 buttonTitle: "Les mer om fordelene ved å handle av et NBF medlem.",
-                linkUrl: URL(string: "https://www.finn.no/nbf.html") ?? URL(string: "https://google.com")!
+                linkUrl: URL(string: "https://google.com")!,
+                isExternal: true
             )
         ),
     ]
